build: Run Progam in Docker Container (#120)
# Pull Request

## Description

This pull request includes several changes to improve the configuration
and setup of the project. The most important changes involve updating
the `.editorconfig` file, adding a new `Dockerfile`, and enhancing the
`Justfile` with Docker commands.

Configuration updates:

*
[`.editorconfig`](diffhunk://#diff-0947e2727d6bad8cd0ac4122f5314bb5b04e337393075bc4b5ef143b17fcbd5bL20-L23):
Removed the configuration for `Justfile` to simplify the file.

Docker setup:

*
[`Dockerfile`](diffhunk://#diff-dd2c0eb6ea5cfc6c4bd4eac30934e2d5746747af48fef6da689e85b752f39557R1-R12):
Added a new `Dockerfile` to set up the Go environment and manage
dependencies, including skipping specific Checkov Docker checks.
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR24-R36):
Added Docker commands to build and run the Docker image for testing the
GitHub action.

Fixes #118