chore: Fix Zizmor Warnings (#133)
# Pull Request

## Description

This pull request includes changes to the permissions configuration in
several GitHub Actions workflow files to improve security and ensure the
correct permissions are granted for specific jobs.

Changes to permissions configuration:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L12-R18):
Removed `statuses` and `security-events` write permissions from the
top-level configuration and added them to the specific jobs that require
them.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L12-R18)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R69-R70)
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL8-R8):
Changed `pull-requests` permission from write to read at the top level
and added write permission to the specific job that requires it.
[[1]](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL8-R8)
[[2]](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR23-R24)
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L12-R24):
Removed `pull-requests` write permission from the top-level
configuration and added it to the specific job that requires it.

Fixes #132