chore: Add Code of Conduct (#137)
# Pull Request

## Description

This pull request includes the addition of a new `CODE_OF_CONDUCT.md`
file to the documentation. The file outlines the expected behavior and
enforcement guidelines for community members to ensure a harassment-free
and inclusive environment.

Key additions include:

* Introduction of a Contributor Covenant Code of Conduct, pledging to
create a welcoming and inclusive community.
* Detailed standards of acceptable and unacceptable behavior within the
community.
* Responsibilities and rights of community leaders in enforcing the Code
of Conduct.
* Scope of the Code of Conduct, applying to all community spaces and
official representations.
* Enforcement guidelines, including consequences ranging from warnings
to permanent bans for violations.

Fixes #138