build: Create Temp Folders (#161)
# Pull Request

## Description

This pull request includes changes to the `main.go` file to enhance the
functionality of the application by adding a new feature for creating
temporary folders. The most important changes are as follows:

New functionality:

*
[`main.go`](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R5-R13):
Added a new function `CreateTempFolders` to create a temporary directory
with subdirectories for the repository's default branch and the commit
being reviewed. This function is now called within the `main` function.
[[1]](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R5-R13)
[[2]](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R42-R64)

Fixes #162