chore: Add CodeQL configuration (#166)
# Pull Request

## Description

This pull request includes updates to the CodeQL configuration and
workflow files to enhance the security and quality checks.

Changes to CodeQL configuration:

*
[`.github/other-configs/codeql-config.yml`](diffhunk://#diff-b868036a8a443102b2e66591b454b6e9f8d7dc04cca30fc4e4dddb7d680d7979R1-R3):
Added a query filter to exclude the `actions/unpinned-tag` query.

Changes to CodeQL workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R84):
Updated the `jobs` section to include the new CodeQL configuration file
for security and quality queries.