chore: Add GitHub Actions CodeQL Scanning (#168)
# Pull Request

## Description

This pull request includes changes to the GitHub Actions workflow
configuration to support multiple languages for CodeQL analysis.

Improvements to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R73-R75):
Added a matrix strategy to run the workflow for both `go` and `actions`
languages.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L82-R85):
Modified the `languages` parameter in the CodeQL initialization step to
use the matrix language.

Fixes #167