chore: Add Project Technologies File (#184)
# Pull Request

## Description

This pull request includes a documentation update to
`docs/PROJECT_TECHNOLOGIES.md`. The change adds a new section that lists
the technologies and frameworks used in the project, organized by
category.

Documentation update:

*
[`docs/PROJECT_TECHNOLOGIES.md`](diffhunk://#diff-070f3e01b55dcf56be3d35ecb65dd331a92f6b31f8cb31a0da4e821b00fe1508R1-R8):
Added a new section that details the technologies and frameworks used in
the project, including Go for backend development and Dependabot and
GitHub Actions for DevOps and CI/CD.

Fixes #86